Remove Committable from state and move ConsensusTime
