+++ crates/hotshot/src/traits/networking/libp2p_network.rs
@@ -25,9 +25,8 @@ use hotshot_types::{
             CommunicationChannel, ConnectedNetwork, ConsensusIntentEvent, FailedToSerializeSnafu,
             NetworkError, NetworkMsg, TestableChannelImplementation, TransmitType, ViewMessage,
         },
-        node_implementation::NodeType,
+        node_implementation::{ConsensusTime,NodeType},
         signature_key::SignatureKey,
-        states::ConsensusTime,
     },
 };
