+++ crates/hotshot/src/traits/storage/memory_storage.rs
@@ -118,7 +118,7 @@ mod test {
         data::{fake_commitment, Leaf},
         simple_certificate::QuorumCertificate,
         traits::{
-            node_implementation::NodeType, signature_key::SignatureKey, states::ConsensusTime,
+            node_implementation::{ConsensusTime,NodeType}, signature_key::SignatureKey,
         },
     };
     use std::marker::PhantomData;
