+++ crates/hotshot/src/types/handle.rs
@@ -26,7 +26,7 @@ use hotshot_types::{
     error::HotShotError,
     event::EventType,
     simple_certificate::QuorumCertificate,
-    traits::{node_implementation::NodeType, states::ConsensusTime, storage::Storage},
+    traits::{node_implementation::{ConsensusTime,NodeType}, storage::Storage},
 };
 use std::sync::Arc;
 use tracing::error;
