+++ crates/task-impls/src/consensus.rs
@@ -28,9 +28,9 @@ use hotshot_types::{
         consensus_api::ConsensusApi,
         election::Membership,
         network::{CommunicationChannel, ConsensusIntentEvent},
-        node_implementation::{NodeImplementation, NodeType},
+        node_implementation::{ConsensusTime,NodeImplementation, NodeType},
         signature_key::SignatureKey,
-        states::{ConsensusTime, ValidatedState},
+        states::{ ValidatedState},
         BlockPayload,
     },
     utils::{Terminator, ViewInner},
