+++ crates/task-impls/src/da.rs
@@ -22,9 +22,8 @@ use hotshot_types::{
         consensus_api::ConsensusApi,
         election::Membership,
         network::{CommunicationChannel, ConsensusIntentEvent},
-        node_implementation::{NodeImplementation, NodeType},
+        node_implementation::{ConsensusTime,NodeImplementation, NodeType},
         signature_key::SignatureKey,
-        states::ConsensusTime,
     },
     utils::ViewInner,
     vote::HasViewNumber,
