+++ crates/testing/src/task_helpers.rs
@@ -25,8 +25,8 @@ use hotshot_types::{
         block_contents::{vid_commitment, BlockHeader, TestableBlock},
         consensus_api::ConsensusApi,
         election::Membership,
-        node_implementation::NodeType,
-        states::{ConsensusTime, ValidatedState},
+        node_implementation::{ConsensusTime,NodeType},
+        states::{ ValidatedState},
         BlockPayload,
     },
     vote::HasViewNumber,
