+++ crates/testing/src/test_runner.rs
@@ -19,7 +19,7 @@ use hotshot_task::{
 use hotshot_types::traits::network::CommunicationChannel;
 use hotshot_types::{
     consensus::ConsensusMetricsValue,
-    traits::{election::Membership, node_implementation::NodeType, states::ConsensusTime},
+    traits::{election::Membership, node_implementation::{ConsensusTime,NodeType}},
     HotShotConfig, ValidatorConfig,
 };
 use std::{
