+++ crates/testing/tests/consensus_task.rs
@@ -12,7 +12,7 @@ use hotshot_types::vote::Certificate;
 use hotshot_types::{
     data::{Leaf, QuorumProposal, ViewNumber},
     message::GeneralConsensusMessage,
-    traits::states::ConsensusTime,
+    traits::node_implementation::ConsensusTime,
 };
 use hotshot_types::{
     simple_vote::QuorumData,
