+++ crates/testing/tests/vid_task.rs
@@ -5,10 +5,10 @@ use hotshot_testing::{
     node_types::{MemoryImpl, TestTypes},
     task_helpers::vid_init,
 };
-use hotshot_types::traits::node_implementation::NodeType;
+use hotshot_types::traits::node_implementation::{ConsensusTime,NodeType};
 use hotshot_types::{
     data::{DAProposal, VidDisperse, VidSchemeTrait, ViewNumber},
-    traits::{consensus_api::ConsensusApi, states::ConsensusTime},
+    traits::{consensus_api::ConsensusApi},
 };
 use std::collections::HashMap;
 use std::marker::PhantomData;
