+++ crates/types/src/simple_certificate.rs
@@ -17,7 +17,7 @@ use crate::{
     },
     traits::{
         election::Membership, node_implementation::NodeType, signature_key::SignatureKey,
-        states::ConsensusTime,
+        node_implementation::ConsensusTime,
     },
     vote::{Certificate, HasViewNumber},
 };
